Upgrading Speed Power Value

+++ src/com/hungryfish/util/FishType.java
@@ -4,17 +4,17 @@ package com.hungryfish.util;
  * User: Breku
  * Date: 24.01.14
  */
-public enum FishType implements Comparable<FishType>{
-    BLACK(4, 3, 4.0f,6),
-    PURPLE(3, 2, 3.0f,5),
-    RED(2, 2, 2.5f,4),
-    YELLOW(1, 2, 2.0f,3),
-    ORANGE(2, 1, 1.5f,2),
-    GREEN(1, 1, 1.0f,1);
+public enum FishType implements Comparable<FishType> {
+    BLACK(4, 3, 4.0f, 6),
+    PURPLE(3, 2, 3.0f, 5),
+    RED(2, 2, 2.5f, 4),
+    YELLOW(1, 2, 2.0f, 3),
+    ORANGE(2, 1, 1.5f, 2),
+    GREEN(1, 1, 1.0f, 1);
 
 
-    private Integer fishValue;
-    private Integer fishPower;
+    private float fishValue;
+    private float fishPower;
     private float fishSpeed;
     private Integer fishLevel;
 
@@ -33,11 +33,11 @@ public enum FishType implements Comparable<FishType>{
         return values()[(ordinal() + 1) % values().length];
     }
 
-    public Integer getFishPower() {
+    public Float getFishPower() {
         return fishPower;
     }
 
-    public Integer getFishValue() {
+    public Float getFishValue() {
         return fishValue;
     }
